backend/requirements.txt
blinker==1.9.0
click==8.1.8
faiss-cpu==1.9.0.post1
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.1
packaging==24.2
Werkzeug==3.1.3
